+++ src/components/tx-flow/flows/ExecuteBatch/ReviewBatch.tsx
@@ -9,7 +9,7 @@ import ErrorMessage from '@/components/tx/ErrorMessage'
 import { ExecutionMethod, ExecutionMethodSelector } from '@/components/tx/ExecutionMethodSelector'
 import DecodedTxs from '@/components/tx-flow/flows/ExecuteBatch/DecodedTxs'
 import SendToBlock from '@/components/tx/SendToBlock'
-import { TxSimulation } from '@/components/tx/NewTxSimulation'
+import { TxSimulation } from '@/components/tx/security/tenderly'
 import { WrongChainWarning } from '@/components/tx/WrongChainWarning'
 import useAsync from '@/hooks/useAsync'
 import { useCurrentChain } from '@/hooks/useChains'
