+++ src/components/tx/SignOrExecuteForm/TxChecks.tsx
@@ -1,8 +1,8 @@
 import { type ReactElement, useContext } from 'react'
-import { TxSimulation } from '../NewTxSimulation'
+import { TxSimulation } from '@/components/tx/security/tenderly'
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
 import { Typography } from '@mui/material'
-import { NewRedefine } from '@/components/tx/security/redefine/NewRedefineScanResult/RedefineScanResult'
+import { Redefine } from '@/components/tx/security/redefine'
 
 const TxChecks = (): ReactElement => {
   const { safeTx } = useContext(SafeTxContext)
@@ -11,9 +11,9 @@ const TxChecks = (): ReactElement => {
     <>
       <Typography variant="h5">Transaction checks</Typography>
 
-      <TxSimulation disabled={false} transactions={safeTx} />
+      <Redefine />
 
-      <NewRedefine />
+      <TxSimulation disabled={false} transactions={safeTx} />
     </>
   )
 }
