--- src/components/tx/TxSimulation/styles.module.css
@@ -1,17 +1,0 @@
-.simulateAccordion :global .MuiAccordionSummary-content {
-  justify-content: space-between;
-  align-items: center;
-}
-
-.loadingText {
-  margin-left: var(--space-1);
-}
-
-.result {
-  border: none;
-}
-
-.skeletonWrapper {
-  border-radius: 8px;
-  overflow: hidden;
-}
