+++ src/components/tx/security/redefine/styles.module.css
@@ -1,0 +1,23 @@
+.hint {
+  padding: var(--space-2);
+  border: unset;
+}
+
+.hint :global .MuiAlert-icon {
+  padding: 0;
+}
+
+.hint :global .MuiAlert-message {
+  padding: 0;
+}
+
+.wrapperBox :global .MuiAccordion-root.Mui-expanded {
+  border-color: var(--color-border-light) !important;
+}
+
+.wrapperBox {
+  border-radius: 6px;
+  border: 1px solid var(--color-border-light);
+  padding: 0;
+  background-color: var(--color-background-main);
+}
