+++ src/components/tx/security/redefine/useRedefine.ts
@@ -12,6 +12,11 @@ import type { SecurityResponse } from '@/services/security/modules/types'
 import { FEATURES } from '@/utils/chains'
 import type { SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import { useState, useEffect, useMemo } from 'react'
+import { type AlertColor, type SvgIconProps } from '@mui/material'
+import { SecuritySeverity } from '@/services/security/modules/types'
+import CloseIcon from '@/public/images/common/close.svg'
+import InfoIcon from '@/public/images/notifications/info.svg'
+import CheckIcon from '@/public/images/common/check.svg'
 
 export const REDEFINE_RETRY_TIMEOUT = 2_000
 const RedefineModuleInstance = new RedefineModule()
@@ -22,6 +27,49 @@ const CRITICAL_ERRORS: Record<number, string> = {
   [1001]: 'Simulation failed',
   [2000]: DEFAULT_ERROR_MESSAGE,
   [3000]: DEFAULT_ERROR_MESSAGE,
+}
+
+type SecurityWarningProps = {
+  color: AlertColor
+  // @ts-expect-error - Use any to avoid conflicts with @svgr/webpack plugin or babel-plugin-inline-react-svg plugin.
+  icon: SvgIconProps['component']
+  label: string
+  action?: string
+}
+
+const ACTION_REJECT = 'Reject this transaction'
+const ACTION_REVIEW = 'Review before processing'
+
+export const mapRedefineSeverity: Record<SecuritySeverity, SecurityWarningProps> = {
+  [SecuritySeverity.CRITICAL]: {
+    action: ACTION_REJECT,
+    color: 'error',
+    icon: CloseIcon,
+    label: 'Critical issue',
+  },
+  [SecuritySeverity.HIGH]: {
+    action: ACTION_REJECT,
+    color: 'error',
+    icon: CloseIcon,
+    label: 'High issue',
+  },
+  [SecuritySeverity.MEDIUM]: {
+    action: ACTION_REVIEW,
+    color: 'warning',
+    icon: InfoIcon,
+    label: 'Medium issue',
+  },
+  [SecuritySeverity.LOW]: {
+    action: ACTION_REVIEW,
+    color: 'warning',
+    icon: InfoIcon,
+    label: 'Low issue',
+  },
+  [SecuritySeverity.NONE]: {
+    color: 'success',
+    icon: CheckIcon,
+    label: 'No issues found',
+  },
 }
 
 export const useRedefine = (
