+++ src/components/tx/security/tenderly/__tests__/utils.test.ts
@@ -1,7 +1,11 @@
 import type { MetaTransactionData, SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import type { SafeInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import { BigNumber, ethers } from 'ethers'
-import { getSimulationPayload, NONCE_STORAGE_POSITION, THRESHOLD_STORAGE_POSITION } from '../utils'
+import {
+  getSimulationPayload,
+  NONCE_STORAGE_POSITION,
+  THRESHOLD_STORAGE_POSITION,
+} from '@/components/tx/security/tenderly/utils'
 import * as safeContracts from '@/services/contracts/safeContracts'
 import { getMultiSendCallOnlyDeployment, getSafeSingletonDeployment } from '@safe-global/safe-deployments'
 import EthSafeTransaction from '@safe-global/safe-core-sdk/dist/src/utils/transactions/SafeTransaction'
