+++ src/components/tx/security/tenderly/styles.module.css
@@ -1,0 +1,4 @@
+.simulate {
+  margin: 0;
+  padding: calc(var(--space-1) / 2) var(--space-2);
+}
